Added code to resume previous slave replication before promoting master
In the event of a network outage:
  Initial: (node1: master, node2: slave)
  node2 -> loses network
  node1 -> writes data to database
  node1 -> loses network
  node1, node2 -> regain network

Before this commit, the node1 database writes would be lost as node1's
master is reset before node2 has a chance to catch up.  With this
commit, node2 will be given an oppurtunity to catch up before node1 is
promoted (and master logs are reset).